Adding daily build and publishing of imgs
Tracking upstream nixpkgs master branch and building the containers off
of this point in time. The idea is that we should be able to have most
up to date set of containers. This can eventually be extended to our
overlay as well, but well start with the container imgs for now.